Remove duplicate `pipeline :browser` declaration
You accidentally added the second one by mistake. I believe the second
one is the only one that will be used, as it will probably override the
first. I did not verify this claim! But the second one also looks like
the one you want, as it has `Auth` plug in it.